+++ src/main/java/gdsc/binaryho/imhere/core/attendance/infrastructure/AttendanceRepository.java
@@ -1,5 +1,6 @@
-package gdsc.binaryho.imhere.core.attendance;
+package gdsc.binaryho.imhere.core.attendance.infrastructure;
 
+import gdsc.binaryho.imhere.core.attendance.Attendance;
 import java.time.LocalDateTime;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
