+++ src/main/java/gdsc/binaryho/imhere/core/lecture/controller/LectureController.java
@@ -2,9 +2,9 @@ package gdsc.binaryho.imhere.core.lecture.controller;
 
 import gdsc.binaryho.imhere.core.attendance.model.response.AttendanceNumberResponse;
 import gdsc.binaryho.imhere.core.lecture.Lecture;
-import gdsc.binaryho.imhere.core.lecture.LectureRepository;
 import gdsc.binaryho.imhere.core.lecture.LectureState;
 import gdsc.binaryho.imhere.core.lecture.application.LectureService;
+import gdsc.binaryho.imhere.core.lecture.infrastructure.LectureRepository;
 import gdsc.binaryho.imhere.core.lecture.model.request.LectureCreateRequest;
 import gdsc.binaryho.imhere.core.lecture.model.response.LectureResponse;
 import io.swagger.v3.oas.annotations.Operation;
