+++ src/main/java/gdsc/binaryho/imhere/core/member/application/MemberService.java
@@ -3,8 +3,8 @@ package gdsc.binaryho.imhere.core.member.application;
 import gdsc.binaryho.imhere.core.auth.exception.MemberNotFoundException;
 import gdsc.binaryho.imhere.core.auth.util.AuthenticationHelper;
 import gdsc.binaryho.imhere.core.member.Member;
-import gdsc.binaryho.imhere.core.member.MemberRepository;
 import gdsc.binaryho.imhere.core.member.Role;
+import gdsc.binaryho.imhere.core.member.infrastructure.MemberRepository;
 import gdsc.binaryho.imhere.core.member.model.request.RoleChangeRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
